@@ -1,26 +1,32 @@
 {
-  "name": "central-factory",
+  "name": "@central-factory/metaverse",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "npm run decorate:patch-package && npm run decorate:angular-cli",
+    "decorate:patch-package": "node_modules/.bin/patch-package",
+    "decorate:angular-cli": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
     "start": "nx serve",
     "build": "nx build",
-    "test": "nx test"
+    "test": "nx test",
+    "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.0",
+    "@angular-architects/module-federation": "^14.0.1",
+    "@angular/animations": "~13.1.0",
     "@angular/cdk": "^13.0.2",
-    "@angular/common": "^12.2.0",
-    "@angular/compiler": "^12.2.0",
-    "@angular/core": "^12.2.0",
+    "@angular/common": "~13.1.0",
+    "@angular/compiler": "~13.1.0",
+    "@angular/core": "~13.1.0",
     "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "^12.2.0",
-    "@angular/platform-browser": "^12.2.0",
-    "@angular/platform-browser-dynamic": "^12.2.0",
-    "@angular/router": "^12.2.0",
+    "@angular/forms": "~13.1.0",
+    "@angular/platform-browser": "~13.1.0",
+    "@angular/platform-browser-dynamic": "~13.1.0",
+    "@angular/router": "~13.1.0",
+    "@asymmetrik/ngx-leaflet": "^8.1.0",
+    "@cliqz/adblocker-electron": "^1.23.2",
     "@docusaurus/core": "^2.0.0-alpha.70",
     "@docusaurus/preset-classic": "^2.0.0-alpha.70",
     "@mdi/svg": "^6.4.95",
@@ -30,63 +36,72 @@
     "@nestjs/platform-express": "^8.2.4",
     "@ng-select/ng-select": "^8.1.1",
     "@ng-stack/forms": "^2.4.0",
-    "@nrwl/angular": "13.1.3",
+    "@nrwl/angular": "13.4.3",
     "clsx": "^1.1.1",
+    "electron-store": "^8.0.1",
     "faker": "^5.5.3",
+    "hotkeys-js": "^3.8.7",
+    "leaflet": "^1.7.1",
     "pouchdb-adapter-idb": "^7.2.2",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "reflect-metadata": "^0.1.13",
     "rxdb": "^10.5.4",
     "rxjs": "~7.5.1",
+    "simple-icons": "^6.5.0",
     "tslib": "^2.0.0",
     "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.0",
-    "@angular-eslint/eslint-plugin": "~12.3.0",
-    "@angular-eslint/eslint-plugin-template": "~12.3.0",
-    "@angular-eslint/template-parser": "~12.3.0",
-    "@angular/cli": "~12.2.0",
-    "@angular/compiler-cli": "^12.2.0",
-    "@angular/language-service": "^12.2.0",
+    "@angular-devkit/build-angular": "~13.1.0",
+    "@angular-eslint/eslint-plugin": "~13.0.1",
+    "@angular-eslint/eslint-plugin-template": "~13.0.1",
+    "@angular-eslint/template-parser": "~13.0.1",
+    "@angular/cli": "~13.1.0",
+    "@angular/compiler-cli": "~13.1.0",
+    "@angular/language-service": "~13.1.0",
     "@compodoc/compodoc": "^1.1.15",
     "@nestjs/schematics": "^8.0.5",
     "@nestjs/testing": "^8.2.4",
-    "@nrwl/cli": "13.1.3",
-    "@nrwl/cypress": "13.1.3",
-    "@nrwl/eslint-plugin-nx": "13.1.3",
-    "@nrwl/jest": "13.1.3",
-    "@nrwl/linter": "13.1.3",
-    "@nrwl/nest": "13.1.3",
-    "@nrwl/node": "13.1.3",
-    "@nrwl/storybook": "13.1.3",
-    "@nrwl/tao": "13.1.3",
-    "@nrwl/workspace": "13.1.3",
+    "@nrwl/cli": "13.4.3",
+    "@nrwl/cypress": "13.4.3",
+    "@nrwl/eslint-plugin-nx": "13.4.3",
+    "@nrwl/jest": "13.4.3",
+    "@nrwl/linter": "13.4.3",
+    "@nrwl/nest": "13.4.3",
+    "@nrwl/node": "13.4.3",
+    "@nrwl/storybook": "13.4.3",
+    "@nrwl/tao": "13.4.3",
+    "@nrwl/workspace": "13.4.3",
     "@nx-plus/docusaurus": "^12.2.0",
-    "@storybook/addon-essentials": "~6.3.0",
-    "@storybook/angular": "~6.3.0",
-    "@storybook/builder-webpack5": "~6.3.0",
-    "@storybook/manager-webpack5": "~6.3.0",
+    "@storybook/addon-essentials": "~6.4.5",
+    "@storybook/angular": "~6.4.5",
+    "@storybook/builder-webpack5": "~6.4.5",
+    "@storybook/manager-webpack5": "~6.4.5",
     "@twittwer/compodoc": "^1.6.4",
     "@types/faker": "^5.5.9",
     "@types/jest": "26.0.24",
+    "@types/leaflet": "^1.7.8",
     "@types/node": "14.14.33",
     "@types/pouchdb-adapter-idb": "^6.1.4",
     "@types/uuid": "^8.3.3",
     "@typescript-eslint/eslint-plugin": "~4.28.3",
     "@typescript-eslint/parser": "~4.28.3",
     "cypress": "^8.3.0",
+    "electron": "^16.0.6",
+    "electron-builder": "^22.14.5",
+    "electron-reload": "^2.0.0-alpha.1",
     "eslint": "7.22.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "27.0.3",
     "jest-preset-angular": "9.0.7",
+    "patch-package": "^6.4.7",
     "prettier": "^2.3.1",
     "style-dictionary": "^3.1.1",
     "ts-jest": "27.0.3",
-    "typescript": "~4.3.5",
+    "typescript": "~4.4.3",
     "yaml": "^1.10.2"
   }
 }
