@@ -1,12 +1,27 @@
 import { Component, Input } from '@angular/core';
 import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
-import type { Application } from '../../../../../models/application';
+import { isElectron } from '@central-factory/web-components/shared/platform/desktop/is-electron';
+import type {
+  Application,
+  ApplicationShortcut,
+} from '../../../../../models/application';
 
 @Component({
   selector: 'cf-application-webview',
   template: `
     <ng-container *ngIf="application">
-      <iframe cfBlock="application-webview" [src]="applicationUrl"></iframe>
+      <iframe
+        *ngIf="!isElectron"
+        cfBlock="application-webview"
+        [src]="applicationUrl"
+      ></iframe>
+      <webview
+        *ngIf="isElectron"
+        cfBlock="application-webview"
+        [attr.src]="applicationUrlString"
+        disablewebsecurity
+        allowpopups
+      ></webview>
     </ng-container>
   `,
 })
@@ -22,9 +37,30 @@ export class ApplicationWebviewComponent {
     return this._application;
   }
 
+  @Input() public set applicationShortcut(
+    value: ApplicationShortcut | undefined
+  ) {
+    this._applicationShortcut = value;
+
+    if (value) {
+      this.applicationUrl = this.applicationShortcut?.url
+        ? this.domSanitizer.bypassSecurityTrustResourceUrl(
+            this.applicationShortcut?.url
+          )
+        : this.applicationUrl;
+    }
+  }
+  public get applicationShortcut(): ApplicationShortcut | undefined {
+    return this._applicationShortcut;
+  }
+
+  public applicationUrlString?: string;
   public applicationUrl?: SafeUrl;
 
+  public isElectron = isElectron();
+
   private _application?: Application;
+  private _applicationShortcut?: ApplicationShortcut;
 
   constructor(private readonly domSanitizer: DomSanitizer) {}
 
@@ -44,6 +80,8 @@ export class ApplicationWebviewComponent {
     if (!applicationDefaultShortcut) {
       throw new Error('Application does not have default shortcut');
     }
+
+    this.applicationUrlString = applicationDefaultShortcut.url;
 
     return this.domSanitizer.bypassSecurityTrustResourceUrl(
       applicationDefaultShortcut.url
