@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { BemModule } from '@central-factory/web-components/angular/bem/bem.module';
 import { ApplicationWebviewComponent } from './application-webview.component';
 
@@ -7,5 +7,6 @@ import { ApplicationWebviewComponent } from './application-webview.component';
   imports: [CommonModule, BemModule],
   declarations: [ApplicationWebviewComponent],
   exports: [ApplicationWebviewComponent],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
 export class ApplicationWebviewModule {}
